requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8
beautifulsoup4>=4.12
python-dotenv>=1.0
gunicorn>=21.0
